requirements.txt
torch>=2.2,<3.0
torchvision>=0.17,<1.0
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
pydantic>=2.6,<3.0
pillow>=10.0,<11.0
numpy>=1.26,<2.0
pytest>=8.0,<9.0
